+++ src/pages/dongmin/Main/Main.scss
@@ -33,130 +33,9 @@ ul {
   text-overflow: ellipsis;
 }
 
-.name-content-cnt {
-  padding-left: 5px;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-
-.size-for-overflow-hidden {
-  width: 300px;
-  height: 15px;
-  margin: auto 0px;
-}
-
-.name {
-  font-size: 10px;
-  font-weight: 700;
-  margin-right: 5px;
-}
-
-.content {
-  font-size: 10px;
-}
-
-.profile-cnt {
-  display: flex;
-  align-items: center;
-  height: 28px;
-  overflow: hidden;
-}
-
-.profile-img {
-  width: 25px;
-  height: 25px;
-  border-radius: 50%;
-  margin-right: 5px;
-}
-
-.profile-text li:nth-child(1) {
-  font-size: 12px;
-  margin: 0;
-}
-.profile-text li:nth-child(2) {
-  font-size: 6px;
-  margin: 0;
-}
-
-@media all and (max-width: 567px) {
+@media all and (max-width: 700px) {
   .main-right {
     display: none;
-  }
-
-  .nav-left {
-    display: none;
-  }
-}
-
-nav {
-  position: sticky;
-  width: 100%;
-  border-bottom: 2px solid rgb(245, 245, 245);
-  padding-top: 10px;
-  padding-bottom: 10px;
-  background-color: white;
-  .nav-cnt {
-    display: flex;
-    justify-content: space-around;
-    .nav-item {
-      display: flex;
-      align-items: center;
-      width: 15px;
-      height: 15px;
-      font-size: 15px;
-      margin: 0;
-    }
-
-    .nav-right img {
-      margin: auto 5px;
-    }
-
-    .nav-left {
-      display: flex;
-      align-items: center;
-      li {
-        display: flex;
-        align-items: center;
-      }
-    }
-
-    .nav-center {
-      border: 1px solid rgb(240, 240, 240);
-      background-color: rgb(250, 250, 250);
-    }
-
-    input {
-      border: none;
-      &::placeholder {
-        opacity: 0.5;
-        text-align: center;
-        vertical-align: center;
-        background: url(../../../assets/dongmin/search.png) no-repeat;
-        background-position: 44px;
-        background-size: 12px;
-        content: '검색';
-      }
-    }
-
-    .nav-right {
-      display: flex;
-      li {
-        display: flex;
-        align-items: center;
-      }
-    }
-
-    .line-vertical {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-    }
-
-    h1 {
-      font-family: 'Lobster', cursive;
-      font-size: 13px;
-    }
   }
 }
 
@@ -166,286 +45,4 @@ main {
   background-color: rgb(250, 250, 250);
   padding-top: 50px;
   overflow: hidden;
-
-  .feeds {
-    padding-bottom: 20px;
-
-    article {
-      width: 300px;
-      border: 2px solid rgb(245, 245, 245);
-      padding-top: 5px;
-      background-color: white;
-    }
-
-    .profile-cnt {
-      display: flex;
-      align-items: center;
-      height: 28px;
-      overflow: hidden;
-    }
-
-    .profile-img {
-      width: 25px;
-      height: 25px;
-      border-radius: 50%;
-      margin-right: 5px;
-    }
-
-    .profile-text li:nth-child(1) {
-      font-size: 12px;
-      margin: 0;
-    }
-
-    .profile-text li:nth-child(2) {
-      font-size: 6px;
-      margin: 0;
-    }
-
-    .img-focus-cnt {
-      width: 300px;
-      height: 300px;
-      overflow: hidden;
-      border: 0;
-      margin: 0;
-      padding: 0;
-    }
-
-    .img-focus {
-      width: 100%;
-      object-fit: cover;
-      margin: 0;
-    }
-
-    button {
-      padding: 0;
-      border: none;
-      background-color: inherit;
-    }
-
-    .button-cnt {
-      display: flex;
-      justify-content: space-between;
-    }
-
-    .icon-cnt {
-      display: flex;
-    }
-
-    .article-icon {
-      width: 15px;
-      height: 15px;
-    }
-
-    .like-cnt {
-      display: flex;
-      align-items: center;
-      font-size: 10px;
-    }
-
-    .like-profile-cnt {
-      display: flex;
-      align-items: center;
-      height: 28px;
-      overflow: hidden;
-    }
-
-    .like-profile-img {
-      width: 18px;
-      height: 18px;
-      border: 2px solid rgb(240, 240, 240);
-      border-radius: 50%;
-      margin-right: 2px;
-    }
-
-    .like-profile-text {
-      font-size: 12px;
-      font-weight: 700;
-      margin: 0;
-    }
-
-    .like-amount-zero {
-      display: none;
-    }
-
-    .like-amount-not-zero {
-      display: relative;
-    }
-
-    hr {
-      margin-top: 5px;
-      opacity: 0.5;
-    }
-
-    .input-cnt {
-      display: flex;
-      justify-content: space-between;
-      margin: 5px auto;
-    }
-
-    .input-cnt input {
-      width: 100%;
-      margin-left: 5px;
-      border: none;
-    }
-
-    .input-cnt button {
-      width: 40px;
-      margin: 0;
-      padding: 0;
-      color: skyblue;
-      border: none;
-    }
-
-    .input-cnt button:hover {
-      color: blue;
-    }
-  }
-}
-
-.main-right {
-  margin-left: 2px;
-  width: 200px;
-  height: 300px;
-  background-color: rgb(250, 250, 250);
-
-  .story-wrap {
-    position: relative;
-    border: 2px solid rgb(245, 245, 245);
-    height: 110px;
-    padding: 8px;
-    background-color: white;
-    font-size: 10px;
-    overflow: scroll;
-
-    &::-webkit-scrollbar {
-      display: none;
-    }
-
-    .show-all {
-      top: 5px;
-      right: 5px;
-      position: absolute;
-      font-size: 10px;
-      font-weight: 700;
-    }
-
-    .story-li {
-      .side-profile-cnt {
-        display: flex;
-        align-items: center;
-        height: 28px;
-        overflow: hidden;
-
-        .side-profile-img {
-          width: 25px;
-          height: 25px;
-          margin-right: 7px;
-          border-radius: 50%;
-        }
-
-        .side-profile-text {
-          display: block;
-          overflow: hidden;
-
-          li {
-            &:nth-child(1) {
-              font-size: 12px;
-              margin: 0;
-            }
-
-            &:nth-child(2) {
-              font-size: 6px;
-              margin: 0;
-              overflow: hidden;
-              white-space: nowrap;
-              text-overflow: ellipsis;
-            }
-          }
-        }
-      }
-    }
-  }
-
-  .recommend-wrap {
-    position: relative;
-    border: 2px solid rgb(245, 245, 245);
-    height: 120px;
-    padding: 8px;
-    background-color: white;
-    font-size: 10px;
-    overflow: scroll;
-
-    &::-webkit-scrollbar {
-      display: none;
-    }
-
-    .show-all {
-      top: 5px;
-      right: 5px;
-      position: absolute;
-      font-size: 10px;
-      font-weight: 700;
-    }
-
-    .story-li {
-      .side-profile-cnt {
-        display: flex;
-        align-items: center;
-        height: 28px;
-        overflow: hidden;
-
-        .side-profile-img {
-          width: 25px;
-          height: 25px;
-          margin-right: 7px;
-          border-radius: 50%;
-        }
-
-        .side-profile-text {
-          display: block;
-          overflow: hidden;
-
-          li {
-            &:nth-child(1) {
-              font-size: 12px;
-              margin: 0;
-            }
-
-            &:nth-child(2) {
-              font-size: 6px;
-              margin: 0;
-              overflow: hidden;
-              white-space: nowrap;
-              text-overflow: ellipsis;
-            }
-          }
-        }
-      }
-    }
-  }
-
-  footer {
-    display: flex;
-    flex-direction: column;
-    margin-top: 30px;
-    opacity: 0.5;
-    li {
-      float: left;
-      margin: 0;
-      font-size: 5px;
-
-      &::after {
-        content: '•';
-      }
-
-      &:last-child::after {
-        content: '';
-      }
-    }
-
-    span {
-      margin-top: 5px;
-      font-size: 7px;
-    }
-  }
 }
