+++ src/pages/dongmin/Main/MainRight/MainRight.js
@@ -1,0 +1,234 @@
+import React from 'react';
+import './MainRight.scss';
+
+const MainRight = () => {
+  return (
+    <div className="main-right">
+      <aside>
+        <div className="profile-cnt">
+          <img
+            className="profile-img"
+            alt="logo"
+            src={process.env.PUBLIC_URL + '/images/dongmin/wecode.png'}
+          />
+          <ul className="profile-text">
+            <li>wecode_bootcamp</li>
+            <li>WeCode - 위코드</li>
+          </ul>
+        </div>
+        <div className="story-wrap">
+          스토리
+          <span className="show-all">모두 보기</span>
+          <ul className="story-li">
+            <li>
+              <div className="side-profile-cnt">
+                <img
+                  className="side-profile-img"
+                  alt="logo"
+                  src={process.env.PUBLIC_URL + '/images/dongmin/junhyun.jpg'}
+                />
+                <ul className="side-profile-text">
+                  <li>junwise.park</li>
+                  <li>16분 전</li>
+                </ul>
+              </div>
+            </li>
+            <li>
+              <div className="side-profile-cnt">
+                <img
+                  className="side-profile-img"
+                  alt="logo"
+                  src={process.env.PUBLIC_URL + '/images/dongmin/junhyun.jpg'}
+                />
+                <ul className="side-profile-text">
+                  <li>junwise.park</li>
+                  <li>3시간 전</li>
+                </ul>
+              </div>
+            </li>
+            <li>
+              <div className="side-profile-cnt">
+                <img
+                  className="side-profile-img"
+                  alt="logo"
+                  src={process.env.PUBLIC_URL + '/images/dongmin/junhyun.jpg'}
+                />
+                <ul className="side-profile-text">
+                  <li>junwise.park</li>
+                  <li>3시간 전</li>
+                </ul>
+              </div>
+            </li>
+            <li>
+              <div className="side-profile-cnt">
+                <img
+                  className="side-profile-img"
+                  alt="logo"
+                  src={process.env.PUBLIC_URL + '/images/dongmin/junhyun.jpg'}
+                />
+                <ul className="side-profile-text">
+                  <li>junwise.park</li>
+                  <li>3시간 전</li>
+                </ul>
+              </div>
+            </li>
+            <li>
+              <div className="side-profile-cnt">
+                <img
+                  className="side-profile-img"
+                  alt="logo"
+                  src={process.env.PUBLIC_URL + '/images/dongmin/junhyun.jpg'}
+                />
+                <ul className="side-profile-text">
+                  <li>junwise.park</li>
+                  <li>6시간 전</li>
+                </ul>
+              </div>
+            </li>
+            <li>
+              <div className="side-profile-cnt">
+                <img
+                  className="side-profile-img"
+                  alt="logo"
+                  src={process.env.PUBLIC_URL + '/images/dongmin/junhyun.jpg'}
+                />
+                <ul className="side-profile-text">
+                  <li>junwise.park</li>
+                  <li>12시간 전</li>
+                </ul>
+              </div>
+            </li>
+          </ul>
+        </div>
+        <div className="recommend-wrap">
+          회원님을 위한 추천
+          <span className="show-all">모두 보기</span>
+          <ul className="story-li">
+            <li>
+              <div className="side-profile-cnt">
+                <img
+                  className="side-profile-img"
+                  alt="logo"
+                  src={process.env.PUBLIC_URL + '/images/dongmin/soohyun.png'}
+                />
+                <ul className="side-profile-text">
+                  <li>watervillage.kim</li>
+                  <li>dongmin.lee 외 2명이 알고 있을지도 모르고 있을지도</li>
+                </ul>
+              </div>
+            </li>
+            <li>
+              <div className="side-profile-cnt">
+                <img
+                  className="side-profile-img"
+                  alt="logo"
+                  src={process.env.PUBLIC_URL + '/images/dongmin/soohyun.png'}
+                />
+                <ul className="side-profile-text">
+                  <li>watervillage.kim</li>
+                  <li>dongmin.lee 외 2명이 알고 있을지도 모르고 있을지도</li>
+                </ul>
+              </div>
+            </li>
+            <li>
+              <div className="side-profile-cnt">
+                <img
+                  className="side-profile-img"
+                  alt="logo"
+                  src={process.env.PUBLIC_URL + '/images/dongmin/soohyun.png'}
+                />
+                <ul className="side-profile-text">
+                  <li>watervillage.kim</li>
+                  <li>dongmin.lee 외 2명이 알고 있을지도 모르고 있을지도</li>
+                </ul>
+              </div>
+            </li>
+            <li>
+              <div className="side-profile-cnt">
+                <img
+                  className="side-profile-img"
+                  alt="logo"
+                  src={process.env.PUBLIC_URL + '/images/dongmin/soohyun.png'}
+                />
+                <ul className="side-profile-text">
+                  <li>watervillage.kim</li>
+                  <li>dongmin.lee 외 2명이 알고 있을지도 모르고 있을지도</li>
+                </ul>
+              </div>
+            </li>
+            <li>
+              <div className="side-profile-cnt">
+                <img
+                  className="side-profile-img"
+                  alt="logo"
+                  src={process.env.PUBLIC_URL + '/images/dongmin/soohyun.png'}
+                />
+                <ul className="side-profile-text">
+                  <li>watervillage.kim</li>
+                  <li>dongmin.lee 외 2명이 알고 있을지도 모르고 있을지도</li>
+                </ul>
+              </div>
+            </li>
+            <li>
+              <div className="side-profile-cnt">
+                <img
+                  className="side-profile-img"
+                  alt="logo"
+                  src={process.env.PUBLIC_URL + '/images/dongmin/soohyun.png'}
+                />
+                <ul className="side-profile-text">
+                  <li>watervillage.kim</li>
+                  <li>dongmin.lee 외 2명이 알고 있을지도 모르고 있을지도</li>
+                </ul>
+              </div>
+            </li>
+            <li>
+              <div className="side-profile-cnt">
+                <img
+                  className="side-profile-img"
+                  alt="logo"
+                  src={process.env.PUBLIC_URL + '/images/dongmin/soohyun.png'}
+                />
+                <ul className="side-profile-text">
+                  <li>watervillage.kim</li>
+                  <li>dongmin.lee 외 2명이 알고 있을지도 모르고 있을지도</li>
+                </ul>
+              </div>
+            </li>
+            <li>
+              <div className="side-profile-cnt">
+                <img
+                  className="side-profile-img"
+                  alt="logo"
+                  src={process.env.PUBLIC_URL + '/images/dongmin/soohyun.png'}
+                />
+                <ul className="side-profile-text">
+                  <li>watervillage.kim</li>
+                  <li>dongmin.lee 외 2명이 알고 있을지도 모르고 있을지도</li>
+                </ul>
+              </div>
+            </li>
+          </ul>
+        </div>
+      </aside>
+      <footer>
+        <ul>
+          <li>Instagram 정보</li>
+          <li>지원</li>
+          <li>홍보 센터</li>
+          <li>API</li>
+          <li>채용 정보</li>
+          <li>개인정보처리방침</li>
+          <li>약관</li>
+          <li>디렉터리</li>
+          <li>프로필</li>
+          <li>해시태그</li>
+          <li>언어</li>
+        </ul>
+        <span>&copy; 2023 INSTAGRAM</span>
+      </footer>
+    </div>
+  );
+};
+
+export default MainRight;
